# src/main/resources/swagger/api.yml
openapi: 3.0.0
info:
  title: Product Inventory
  description: The Product Inventory Management API provides standardized mechanism for
    product inventory management such as creation, update and retrieval of the
    representation of a product in the inventory. It also allows the
    notification of events related to product lifecycle. A product offering
    procured by a customer or other interested party playing a party role. A
    product is realized as one or more service(s) and / or resource(s).
  version: 4.0.0
...
paths:
  /product:
    get:
      operationId: listProduct
      summary: List or find Product objects
      description: This operation list or find Product entities
      tags:
        - product
      parameters:
        - name: fields
          description: Comma-separated properties to be provided in response
          required: false
          in: query
          schema:
            type: string
        - name: offset
          description: Requested index for start of resources to be provided in response
          required: false
          in: query
          schema:
            type: integer
        - name: limit
          description: Requested number of resources to be provided in response
          required: false
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Success
          headers:
            X-Result-Count:
              description: Actual number of items returned in the response body
              schema:
                type: integer
            X-Total-Count:
              description: Total number of items matching criteria
              schema:
                type: integer
          content:
            application/json;charset=utf-8:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not Found
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createProduct
      summary: Creates a Product
      description: This operation creates a Product entity.
      tags:
        - product
      requestBody:
        content:
          application/json;charset=utf-8:
            schema:
              $ref: "#/components/schemas/Product_Create"
        description: The Product to be created
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Bad Request
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Forbidden
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "405":
          description: Method Not allowed
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Conflict
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal Server Error
          content:
            application/json;charset=utf-8:
              schema:
                $ref: "#/components/schemas/Error"
...
servers:
  - url: https://serverRoot//api/productInventory/v4
components:
  schemas:
    AgreementItemRef:
      type: object
      description: Agreement reference. An agreement represents a contract or arrangement,
        either written or verbal and sometimes enforceable by law, such as a
        service level agreement or a customer price agreement. An agreement
        involves a number of other business entities, such as products,
        services, and resources and/or their specifications.
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        agreementItemId:
          type: string
          description: Identifier of the agreement
        name:
          type: string
          description: Name of the related entity.
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    Any: {}
    BillingAccountRef:
      type: object
      description: BillingAccount reference. A BillingAccount is a detailed description of
        a bill structure.
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of the billing account
        href:
          type: string
          description: Reference of the billing account
        name:
          type: string
          description: Name of the billing account
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    Characteristic:
      type: object
      description: Describes a given characteristic of an object or entity through a
        name/value pair.
      required:
        - name
        - value
      properties:
        name:
          type: string
          description: Name of the characteristic
        valueType:
          type: string
          description: Data type of the value of the characteristic
        value:
          $ref: "#/components/schemas/Any"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
...
    Product:
      type: object
      description: A product offering procured by a customer or other interested party
        playing a party role. A product is realized as one or more service(s)
        and / or resource(s).
      properties:
        id:
          type: string
          description: Unique identifier of the product
        href:
          type: string
          description: Reference of the product
        description:
          type: string
          description: Is the description of the product. It could be copied from the
            description of the Product Offering.
        isBundle:
          type: boolean
          description: If true, the product is a ProductBundle which is an instantiation of
            a BundledProductOffering. If false, the product is a
            ProductComponent which is an instantiation of a
            SimpleProductOffering.
        isCustomerVisible:
          type: boolean
          description: If true, the product is visible by the customer.
        name:
          type: string
          description: Name of the product. It could be the same as the name of the product
            offering
        orderDate:
          type: string
          format: date-time
          description: Is the date when the product was ordered
        productSerialNumber:
          type: string
          description: Is the serial number for the product. This is typically applicable
            to tangible products e.g. Broadband Router.
        startDate:
          type: string
          format: date-time
          description: Is the date from which the product starts
        terminationDate:
          type: string
          format: date-time
          description: Is the date when the product was terminated
        agreement:
          type: array
          items:
            $ref: "#/components/schemas/AgreementItemRef"
        billingAccount:
          $ref: "#/components/schemas/BillingAccountRef"
        place:
          type: array
          items:
            $ref: "#/components/schemas/RelatedPlaceRefOrValue"
        product:
          type: array
          items:
            $ref: "#/components/schemas/ProductRefOrValue"
        productCharacteristic:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        productOffering:
          $ref: "#/components/schemas/ProductOfferingRef"
        productOrderItem:
          type: array
          items:
            $ref: "#/components/schemas/RelatedProductOrderItem"
        productPrice:
          type: array
          items:
            $ref: "#/components/schemas/ProductPrice"
        productRelationship:
          type: array
          items:
            $ref: "#/components/schemas/ProductRelationship"
        productSpecification:
          $ref: "#/components/schemas/ProductSpecificationRef"
        productTerm:
          type: array
          items:
            $ref: "#/components/schemas/ProductTerm"
        realizingResource:
          type: array
          items:
            $ref: "#/components/schemas/ResourceRef"
        realizingService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRef"
        relatedParty:
          type: array
          items:
            $ref: "#/components/schemas/RelatedParty"
        status:
          $ref: "#/components/schemas/ProductStatusType"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
    Product_Create:
      type: object
      description: >-
        A product offering procured by a customer or other interested party
        playing a party role. A product is realized as one or more service(s)
        and / or resource(s).

        Skipped properties: id,href
      required:
        - status
      properties:
        description:
          type: string
          description: Is the description of the product. It could be copied from the
            description of the Product Offering.
        isBundle:
          type: boolean
          description: If true, the product is a ProductBundle which is an instantiation of
            a BundledProductOffering. If false, the product is a
            ProductComponent which is an instantiation of a
            SimpleProductOffering.
        isCustomerVisible:
          type: boolean
          description: If true, the product is visible by the customer.
        name:
          type: string
          description: Name of the product. It could be the same as the name of the product
            offering
        orderDate:
          type: string
          format: date-time
          description: Is the date when the product was ordered
        productSerialNumber:
          type: string
          description: Is the serial number for the product. This is typically applicable
            to tangible products e.g. Broadband Router.
        startDate:
          type: string
          format: date-time
          description: Is the date from which the product starts
        terminationDate:
          type: string
          format: date-time
          description: Is the date when the product was terminated
        agreement:
          type: array
          items:
            $ref: "#/components/schemas/AgreementItemRef"
        billingAccount:
          $ref: "#/components/schemas/BillingAccountRef"
        place:
          type: array
          items:
            $ref: "#/components/schemas/RelatedPlaceRefOrValue"
        product:
          type: array
          items:
            $ref: "#/components/schemas/ProductRefOrValue"
        productCharacteristic:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        productOffering:
          $ref: "#/components/schemas/ProductOfferingRef"
        productOrderItem:
          type: array
          items:
            $ref: "#/components/schemas/RelatedProductOrderItem"
        productPrice:
          type: array
          items:
            $ref: "#/components/schemas/ProductPrice"
        productRelationship:
          type: array
          items:
            $ref: "#/components/schemas/ProductRelationship"
        productSpecification:
          $ref: "#/components/schemas/ProductSpecificationRef"
        productTerm:
          type: array
          items:
            $ref: "#/components/schemas/ProductTerm"
        realizingResource:
          type: array
          items:
            $ref: "#/components/schemas/ResourceRef"
        realizingService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRef"
        relatedParty:
          type: array
          items:
            $ref: "#/components/schemas/RelatedParty"
        status:
          $ref: "#/components/schemas/ProductStatusType"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
...
    ProductOfferingRef:
      type: object
      description: ProductOffering reference. A product offering represents entities that
        are orderable from the provider of the catalog, this resource includes
        pricing information.
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        name:
          type: string
          description: Name of the related entity.
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    ProductPrice:
      type: object
      description: An amount, usually of money, that represents the actual price paid by a
        Customer for a purchase, a rent or a lease of a Product. The price is
        valid for a defined period of time.
      required:
        - price
        - priceType
      properties:
        description:
          type: string
          description: A narrative that explains in detail the semantics of this product
            price.
        name:
          type: string
          description: A short descriptive name such as "Subscription price".
        priceType:
          type: string
          description: A category that describes the price, such as recurring, discount,
            allowance, penalty, and so forth.
        recurringChargePeriod:
          type: string
          description: Could be month, week...
        unitOfMeasure:
          type: string
          description: Could be minutes, GB...
        billingAccount:
          $ref: "#/components/schemas/BillingAccountRef"
        price:
          $ref: "#/components/schemas/Price"
        productOfferingPrice:
          $ref: "#/components/schemas/ProductOfferingPriceRef"
        productPriceAlteration:
          type: array
          items:
            $ref: "#/components/schemas/PriceAlteration"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
...
    ProductRefOrValue:
      type: object
      description: A product to be created defined by value or existing defined by
        reference. The polymorphic attributes @type, @schemaLocation &
        @referredType are related to the product entity and not the
        RelatedProductRefOrValue class itself
      properties:
        id:
          type: string
          description: Unique identifier of the product
        href:
          type: string
          description: Reference of the product
        description:
          type: string
          description: Is the description of the product. It could be copied from the
            description of the Product Offering.
        isBundle:
          type: boolean
          description: If true, the product is a ProductBundle which is an instantiation of
            a BundledProductOffering. If false, the product is a
            ProductComponent which is an instantiation of a
            SimpleProductOffering.
        isCustomerVisible:
          type: boolean
          description: If true, the product is visible by the customer.
        name:
          type: string
          description: Name of the product. It could be the same as the name of the product
            offering
        orderDate:
          type: string
          format: date-time
          description: Is the date when the product was ordered
        productSerialNumber:
          type: string
          description: Is the serial number for the product. This is typically applicable
            to tangible products e.g. Broadband Router.
        startDate:
          type: string
          format: date-time
          description: Is the date from which the product starts
        terminationDate:
          type: string
          format: date-time
          description: Is the date when the product was terminated
        agreement:
          type: array
          items:
            $ref: "#/components/schemas/AgreementItemRef"
        billingAccount:
          $ref: "#/components/schemas/BillingAccountRef"
        place:
          type: array
          items:
            $ref: "#/components/schemas/RelatedPlaceRefOrValue"
        product:
          type: array
          items:
            $ref: "#/components/schemas/ProductRefOrValue"
        productCharacteristic:
          type: array
          items:
            $ref: "#/components/schemas/Characteristic"
        productOffering:
          $ref: "#/components/schemas/ProductOfferingRef"
        productOrderItem:
          type: array
          items:
            $ref: "#/components/schemas/RelatedProductOrderItem"
        productPrice:
          type: array
          items:
            $ref: "#/components/schemas/ProductPrice"
        productRelationship:
          type: array
          items:
            $ref: "#/components/schemas/ProductRelationship"
        productSpecification:
          $ref: "#/components/schemas/ProductSpecificationRef"
        productTerm:
          type: array
          items:
            $ref: "#/components/schemas/ProductTerm"
        realizingResource:
          type: array
          items:
            $ref: "#/components/schemas/ResourceRef"
        realizingService:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRef"
        relatedParty:
          type: array
          items:
            $ref: "#/components/schemas/RelatedParty"
        status:
          $ref: "#/components/schemas/ProductStatusType"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    ProductRelationship:
      type: object
      description: Linked products to the one instantiate, such as [bundled] if the product
        is a bundle and you want to describe the bundled products inside this
        bundle; [reliesOn] if the product needs another already owned product to
        rely on (e.g. an option on an already owned mobile access product)
        [targets] or [isTargeted] (depending on the way of expressing the link)
        for any other kind of links that may be useful
      required:
        - product
        - relationshipType
      properties:
        relationshipType:
          type: string
          description: Type of the product relationship, such as [bundled] if the product
            is a bundle and you want to describe the bundled products inside
            this bundle; [reliesOn] if the product needs another already owned
            product to rely on (e.g. an option on an already owned mobile access
            product) [targets] or [isTargeted] (depending on the way of
            expressing the link) for any other kind of links that may be useful
        product:
          $ref: "#/components/schemas/ProductRefOrValue"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
    ProductSpecificationRef:
      type: object
      description: "Product specification reference: A ProductSpecification is a detailed
        description of a tangible or intangible object made available externally
        in the form of a ProductOffering to customers or other parties playing a
        party role."
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        name:
          type: string
          description: Name of the related entity.
        version:
          type: string
          description: Version of the product specification
        targetProductSchema:
          $ref: "#/components/schemas/TargetProductSchema"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
...
    ProductTerm:
      type: object
      description: Description of a productTerm linked to this product. This represent a
        commitment with a duration
      properties:
        description:
          type: string
          description: Description of the productTerm
        name:
          type: string
          description: Name of the productTerm
        duration:
          $ref: "#/components/schemas/Quantity"
        validFor:
          $ref: "#/components/schemas/TimePeriod"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
...
    RelatedParty:
      type: object
      description: Related Entity reference. A related party defines party or party role
        linked to a specific entity.
      required:
        - "@referredType"
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        name:
          type: string
          description: Name of the related entity.
        role:
          type: string
          description: Role played by the related party
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    RelatedPlaceRefOrValue:
      type: object
      description: Related Entity reference. A related place defines a place described by
        reference or by value linked to a specific entity. The polymorphic
        attributes @type, @schemaLocation & @referredType are related to the
        place entity and not the RelatedPlaceRefOrValue class itself
      required:
        - role
      properties:
        id:
          type: string
          description: Unique identifier of the place
        href:
          type: string
          description: Unique reference of the place
        name:
          type: string
          description: A user-friendly name for the place, such as [Paris Store], [London
            Store], [Main Home]
        role:
          type: string
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    RelatedProductOrderItem:
      type: object
      description: RelatedProductOrderItem (ProductOrder item) .The product order item
        which triggered product creation/change/termination.
      required:
        - orderItemId
        - productOrderId
      properties:
        orderItemAction:
          type: string
          description: Action of the order item for this product
        orderItemId:
          type: string
          description: Identifier of the order item where the product was managed
        productOrderHref:
          type: string
          description: Reference of the related entity.
        productOrderId:
          type: string
          description: Unique identifier of a related entity.
        role:
          type: string
          description: role of the product order item for this product
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    ResourceRef:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        name:
          type: string
          description: Name of the resource
        value:
          type: string
          description: "The resource value that can be used to identify a resource with a
            public key (e.g.: a tel nr, an msisdn)"
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
    ServiceRef:
      type: object
      description: Service reference, for when Service is used by other entities
      required:
        - id
      properties:
        id:
          type: string
          description: Unique identifier of a related entity.
        href:
          type: string
          description: Reference of the related entity.
        name:
          type: string
          description: Name of the related entity.
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name
        "@referredType":
          type: string
          description: The actual type of the target instance when needed for
            disambiguation.
...
    Error:
      description: Used when an API throws an Error, typically with a HTTP error
        response-code (3xx, 4xx, 5xx)
      type: object
      required:
        - code
        - reason
      properties:
        code:
          type: string
          description: Application relevant detail, defined in the API or a common list.
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown to a
            client user.
        message:
          type: string
          description: More details and corrective actions related to the error which can
            be shown to a client user.
        status:
          type: string
          description: HTTP Error code extension
        referenceError:
          type: string
          format: uri
          description: URI of documentation describing the error.
        "@baseType":
          type: string
          description: When sub-classing, this defines the super-class.
        "@schemaLocation":
          type: string
          format: uri
          description: A URI to a JSON-Schema file that defines additional attributes and
            relationships
        "@type":
          type: string
          description: When sub-classing, this defines the sub-class entity name.